Remove need for to_cudf_compatible_scalar
